synchrotrace: refactored the core runtime.
Components were previously highly coupled,
thus it was deemed more worthwhile to refactor all the
components together than incrementally refactor.

Decoupling
----------

The 'SynchroTrace' runtime has been decupled into
 - a trace parser
 - a replayer.

Parsing
-------

Refactored the parsing to be decoupled from replay.
Additionally made parsing more efficient with less strings being generated.

The new parser also requires a new trace format that is faster to parse.
Unique symbols are located at the beginning of each event to be able
to quickly detect the type of event. This means existing traces cannot
be used. There is current work on a tool that can convert old traces to
new traces.

Event Generation
----------------

- Memory address generation

The SynchroTrace replayer does not have a TLB.
In lieu of this, it converts the virtual addresses in the trace to
physical
addresses when memory events are generated from the trace.
The following algorithm was used:

    # remove offset bits from the largest phys memory address
    largest_tag_index = main_memory_size / cache_block_size

    # constrain virtual addr to fit within phys memory
    # while keeping the same location in the cache block
    vaddr_tag_index = vaddr / cache_block_size
    vaddr_tag_index_modified = vaddr_tag_index % largest_tag_index
    paddr = (vaddr_tag_index_modified * cache_block_size) | offset

This algorithm modulos the virtual address to fit within physical
memory, while keeping its position within the cache block.
However, this computation is needlessly redundant if physical memory is
a multiple of the cache block size, which is almost always the case.
If main memory is a multiple of cache block size, then the above
algorithm is effectively:

    paddr = vaddr % main_memory_size

Instead, a constraint was added to ensure physical memory is a multiple
of cache block size, and the algorithm was simplified:

    assert(isPowerOf2(cache_block_size) &&
           (main_memory_size % cache_block_size == 0))
    paddr = vaddr % main_memory_size

This still accomodates a large variety of likely main memory ranges.

- Events

Events are now simply tagged unions.
The previous implementation had events being a *combination* of 1 or
more types, instead of a simpler sum type (only 1 event at a time).

This simplifies replay, which can more easily handle one event at time.
Each replay handler chooses to consume the event, reschedule the core,
and swap threads.

Synchronization Replay
----------------------

Modify synchronization handling to be more explicit.
Heuristics handling signal and broadcast are the same as before.

The biggest differences functional change is that threads are now
only swapped when a thread is blocked, instead of the previous
implementation that forced a thread swap attempt after every *trace event*
(a line in the trace file) completed.
A reason for this could not be inferred, other than a heuristic to prevent
deadlocks, or a rudimentary kernel scheduler.
In the case of a scheduler, the context swap cost was so small it would not
be sufficient to represent an actual scheduler.

The previous behavior can be restored if the previous motivations become
clear.

Change-Id: Ie01de379a2a49b9ea684b90ba494bf7908cc2f15