synchrotrace/replay: fix double scheduled core
Events are only allowed to be scheduled once. After the event
fires, it can be scheduled again. When a new thread is created,
it may be mapped to a core that's already scheduled with an
active thread, so check if the new thread's core is scheduled
and, if it's not, then it can be scheduled.

Change-Id: Ib4e49fe30e64a3a77cb67196c129d90ffa3e5d7c