synchrotrace/replay: context swap caller reschedules self
Change `tryCxtSwapAndSchedule` to only inform the caller if the context swap
is unsuccessful, instead of automatically rescheduling the calling core/thread.
It is then up to the caller to decide how to handle the fact that it is still
considered the "active" thread the core will run. Generally, the thread will
reschedule itself to check at a future tick whether it is still blocked or not.

Change-Id: I265d4745b9f91bf6ce0504753cbfb11cb3fc9662